pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nitrogen"
version = "1.0.0"
description = "VLM-based game playing agent"
readme = "README.md"
requires-python = ">=3.10"

# Default: everything
dependencies = [
    # Shared
    "numpy",
    "pyzmq",
    
    # Serve
    "torch",
    "pyyaml",
    "einops",
    "transformers",
    "pydantic",
    "diffusers",
    "polars",
    
    # Play (Windows-only deps marked)
    "pillow",
    "opencv-python",
    "gymnasium",
    "av",
]

[project.optional-dependencies]
serve = [
    "numpy",
    "pyzmq",
    "torch",
    "pyyaml",
    "einops",
    "transformers",
    "pydantic",
    "diffusers",
    "polars",
]

play = [
    "numpy",
    "pyzmq",
    "pillow",
    "opencv-python",
    "gymnasium",
    "av",
]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["scripts*"]
